chore: configure policy in project level
