+++ frontend/src/components/GeneralSetting/SecuritySetting.vue
@@ -82,37 +82,11 @@
           {{ $t("settings.general.workspace.require-2fa.description") }}
         </div>
       </div>
-      <div class="mb-7 mt-4 lg:mt-0">
-        <NTooltip placement="top-start" :disabled="allowEdit">
-          <template #trigger>
-            <label
-              class="flex items-center gap-x-2"
-              :class="[allowEdit ? 'cursor-pointer' : 'cursor-not-allowed']"
-            >
-              <NCheckbox
-                :disabled="!allowEdit"
-                :checked="restrictIssueCreationForSQLReview"
-                :label="
-                  $t(
-                    'settings.general.workspace.restrict-issue-creation-for-sql-review.title'
-                  )
-                "
-                @update:checked="handleRestrictIssueCreationForSQLReviewToggle"
-              />
-            </label>
-          </template>
-          <span class="text-sm text-gray-400 -translate-y-2">
-            {{ $t("settings.general.workspace.only-admin-can-edit") }}
-          </span>
-        </NTooltip>
-        <div class="mb-3 text-sm text-gray-400">
-          {{
-            $t(
-              "settings.general.workspace.restrict-issue-creation-for-sql-review.description"
-            )
-          }}
-        </div>
-      </div>
+      <RestrictIssueCreationConfigure
+        class="mb-7 mt-4 lg:mt-0"
+        :resource="''"
+        :allow-edit="allowEdit"
+      />
       <SignInFrequencySetting :allow-edit="allowEdit" />
       <MaximumRoleExpirationSetting :allow-edit="allowEdit" />
       <DomainRestrictionSetting :allow-edit="allowEdit" />
@@ -129,20 +103,11 @@
 <script lang="ts" setup>
 import { NCheckbox } from "naive-ui";
 import { storeToRefs } from "pinia";
-import { computed, onMounted, reactive } from "vue";
+import { computed, reactive } from "vue";
 import { useI18n } from "vue-i18n";
-import {
-  featureToRef,
-  pushNotification,
-  useActuatorV1Store,
-  usePolicyV1Store,
-} from "@/store";
+import { featureToRef, pushNotification, useActuatorV1Store } from "@/store";
 import { useSettingV1Store } from "@/store/modules/v1/setting";
 import type { FeatureType } from "@/types";
-import {
-  PolicyResourceType,
-  PolicyType,
-} from "@/types/proto/v1/org_policy_service";
 import DomainRestrictionSetting from "./DomainRestrictionSetting.vue";
 import MaximumRoleExpirationSetting from "./MaximumRoleExpirationSetting.vue";
 import SignInFrequencySetting from "./SignInFrequencySetting.vue";
@@ -159,7 +124,6 @@ const state = reactive<LocalState>({});
 const { t } = useI18n();
 const settingV1Store = useSettingV1Store();
 const actuatorStore = useActuatorV1Store();
-const policyV1Store = usePolicyV1Store();
 
 const { isSaaSMode } = storeToRefs(actuatorStore);
 const hasWatermarkFeature = featureToRef("bb.feature.branding");
@@ -178,25 +142,6 @@ const disallowSignupEnabled = computed((): boolean => {
 const require2FAEnabled = computed((): boolean => {
   return settingV1Store.workspaceProfileSetting?.require2fa ?? false;
 });
-const restrictIssueCreationForSQLReview = computed((): boolean => {
-  return (
-    policyV1Store.getPolicyByParentAndType({
-      parentPath: "",
-      policyType: PolicyType.RESTRICT_ISSUE_CREATION_FOR_SQL_REVIEW,
-    })?.restrictIssueCreationForSqlReviewPolicy?.disallow ?? false
-  );
-});
-
-onMounted(async () => {
-  await prepareOrgPolicy();
-});
-
-const prepareOrgPolicy = async () => {
-  await policyV1Store.getOrFetchPolicyByParentAndType({
-    parentPath: "",
-    policyType: PolicyType.RESTRICT_ISSUE_CREATION_FOR_SQL_REVIEW,
-  });
-};
 
 const handleDisallowSignupToggle = async (on: boolean) => {
   if (!hasDisallowSignupFeature.value && on) {
@@ -251,22 +196,6 @@ const handleWatermarkToggle = async (on: boolean) => {
     module: "bytebase",
     style: "SUCCESS",
     title: t("settings.general.workspace.watermark.update-success"),
-  });
-};
-
-const handleRestrictIssueCreationForSQLReviewToggle = async (on: boolean) => {
-  await policyV1Store.createPolicy("", {
-    type: PolicyType.RESTRICT_ISSUE_CREATION_FOR_SQL_REVIEW,
-    resourceType: PolicyResourceType.WORKSPACE,
-    restrictIssueCreationForSqlReviewPolicy: {
-      disallow: on,
-    },
-  });
-
-  pushNotification({
-    module: "bytebase",
-    style: "SUCCESS",
-    title: t("settings.general.workspace.config-updated"),
   });
 };
 </script>
